requirements.txt
fake-bpy-module-latest
ruff
